Major refactor: no code changes just repo restructure
