+++ server/constants.py
@@ -1,0 +1,15 @@
+ERROR_CODES = {
+    'INVALID_PARAMETER': -32603,
+    'TOOL_EXECUTION_FAILED': -32604,
+    'INTERNAL_ERROR': -32603,
+    'PAYLOAD_TOO_LARGE': -32600,
+    'PAYLOAD_TOO_COMPLEX': -32600
+}
+
+PAYLOAD_LIMITS = {
+    'MAX_PAYLOAD_SIZE': 5_242_880,
+    'MAX_JSON_DEPTH': 20,
+    'MAX_OBJECT_KEYS': 500,
+    'MAX_ARRAY_LENGTH': 2000,
+    'MAX_STRING_LENGTH': 500_000
+}
