+++ server/tools.py
@@ -1,139 +1,3 @@
-#!/usr/bin/env python3
-
-import sys
-import os
-import json
-from http.server import BaseHTTPRequestHandler
-import asyncio
-from typing import Dict, Any
-import traceback
-
-# Import required modules for serverless environment
-import uuid
-import hashlib
-
-# Enhanced error codes following MCP standards
-ERROR_CODES = {
-    'INVALID_PARAMETER': -32603,
-    'TOOL_EXECUTION_FAILED': -32604,
-    'INTERNAL_ERROR': -32603,
-    'PAYLOAD_TOO_LARGE': -32600,
-    'PAYLOAD_TOO_COMPLEX': -32600
-}
-
-# Payload validation limits
-# These limits balance security (DoS prevention) with practical MCP usage
-PAYLOAD_LIMITS = {
-    'MAX_PAYLOAD_SIZE': 5_242_880,  # 5MB max payload size (increased for large threat models)
-    'MAX_JSON_DEPTH': 20,            # Maximum nesting depth (MCP protocol + nested threat data)
-    'MAX_OBJECT_KEYS': 500,          # Maximum keys in single object (rich threat metadata)
-    'MAX_ARRAY_LENGTH': 2000,        # Maximum array length (large-scale threat assessments)
-    'MAX_STRING_LENGTH': 500_000     # Maximum string length (500KB - detailed descriptions)
-}
-
-# Simplified tool implementations for Vercel deployment
-# Note: These provide framework and guidance for LLM client analysis
-
-def sanitize_error(error: Exception, error_context: str = "") -> tuple[str, str]:
-    """
-    Sanitize error messages to prevent information disclosure.
-
-    Returns:
-        tuple: (error_id, sanitized_message) where:
-            - error_id: Unique identifier for correlating with server logs
-            - sanitized_message: Generic error message safe for client
-
-    Security measures:
-        - Logs full error details to server logs (stderr)
-        - Returns generic message to client
-        - Generates unique error ID for correlation
-        - Prevents leakage of: stack traces, file paths, internal implementation details
-    """
-    # Generate unique error ID for correlation
-    error_id = str(uuid.uuid4())[:8]
-
-    # Log detailed error information to server logs (stderr goes to Vercel logs)
-    print(f"[ERROR {error_id}] Context: {error_context}", file=sys.stderr)
-    print(f"[ERROR {error_id}] Exception Type: {type(error).__name__}", file=sys.stderr)
-    print(f"[ERROR {error_id}] Exception Message: {str(error)}", file=sys.stderr)
-    print(f"[ERROR {error_id}] Stack Trace:", file=sys.stderr)
-    traceback.print_exc(file=sys.stderr)
-
-    # Return generic message safe for client
-    sanitized_message = f"An internal error occurred. Error ID: {error_id}"
-
-    return error_id, sanitized_message
-
-def validate_json_complexity(data: Any, current_depth: int = 0) -> Dict[str, Any]:
-    """
-    Recursively validate JSON complexity to prevent DoS attacks.
-
-    Checks:
-    - Maximum nesting depth
-    - Maximum number of keys in objects
-    - Maximum array length
-    - Maximum string length
-
-    Returns:
-        Dict with 'valid' (bool) and 'error' (str) if invalid
-    """
-    # Check depth limit
-    if current_depth > PAYLOAD_LIMITS['MAX_JSON_DEPTH']:
-        return {
-            'valid': False,
-            'error': f"JSON nesting depth exceeds maximum of {PAYLOAD_LIMITS['MAX_JSON_DEPTH']}"
-        }
-
-    # Validate dictionaries/objects
-    if isinstance(data, dict):
-        # Check number of keys
-        if len(data) > PAYLOAD_LIMITS['MAX_OBJECT_KEYS']:
-            return {
-                'valid': False,
-                'error': f"Object contains {len(data)} keys, exceeds maximum of {PAYLOAD_LIMITS['MAX_OBJECT_KEYS']}"
-            }
-
-        # Recursively validate values
-        for key, value in data.items():
-            # Validate key length
-            if isinstance(key, str) and len(key) > PAYLOAD_LIMITS['MAX_STRING_LENGTH']:
-                return {
-                    'valid': False,
-                    'error': f"Object key length exceeds maximum of {PAYLOAD_LIMITS['MAX_STRING_LENGTH']}"
-                }
-
-            # Recursively validate value
-            result = validate_json_complexity(value, current_depth + 1)
-            if not result['valid']:
-                return result
-
-    # Validate arrays
-    elif isinstance(data, list):
-        # Check array length
-        if len(data) > PAYLOAD_LIMITS['MAX_ARRAY_LENGTH']:
-            return {
-                'valid': False,
-                'error': f"Array length {len(data)} exceeds maximum of {PAYLOAD_LIMITS['MAX_ARRAY_LENGTH']}"
-            }
-
-        # Recursively validate elements
-        for item in data:
-            result = validate_json_complexity(item, current_depth + 1)
-            if not result['valid']:
-                return result
-
-    # Validate strings
-    elif isinstance(data, str):
-        if len(data) > PAYLOAD_LIMITS['MAX_STRING_LENGTH']:
-            return {
-                'valid': False,
-                'error': f"String length {len(data)} exceeds maximum of {PAYLOAD_LIMITS['MAX_STRING_LENGTH']}"
-            }
-
-    # Other types (int, float, bool, None) are inherently safe
-
-    return {'valid': True, 'error': None}
-
 
 def get_stride_threat_framework(args: Dict[str, Any]) -> Dict[str, Any]:
     """Provide STRIDE threat modeling framework for LLM client analysis."""
@@ -1463,558 +1327,3 @@ Don't aim for perfect information - aim for sufficient information to identify t
     base_framework["repository_context"] = repo_context
 
     return base_framework
-
-def handle_mcp_request(body: dict) -> dict:
-    """Handle MCP JSON-RPC requests using the improved MCP server"""
-    
-    method = body.get('method')
-    params = body.get('params', {})
-    request_id = body.get('id')
-    
-    
-    # Handle initialize
-    if method == 'initialize':
-        return {
-            "jsonrpc": "2.0",
-            "result": {
-                "protocolVersion": "2025-03-26",
-                "capabilities": {
-                    "tools": {"listChanged": False}
-                },
-                "serverInfo": {
-                    "name": "STRIDE GPT MCP Server",
-                    "version": "0.1.0"
-                },
-                "instructions": "Professional threat modeling server using the STRIDE methodology."
-            },
-            "id": request_id
-        }
-    
-    # Handle tools/list
-    elif method == 'tools/list':
-        tools = [
-            {
-                "name": "get_stride_threat_framework",
-                "description": "Get comprehensive STRIDE threat modeling framework and guidance for threat analysis",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "app_description": {
-                            "type": "string",
-                            "description": "Detailed description of the application architecture and functionality"
-                        },
-                        "app_type": {
-                            "type": "string", 
-                            "description": "Type of application",
-                            "default": "Web Application"
-                        },
-                        "authentication_methods": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "List of authentication methods used",
-                            "default": ["Username/Password"]
-                        },
-                        "internet_facing": {
-                            "type": "boolean",
-                            "description": "Whether the application is accessible from the internet",
-                            "default": True
-                        },
-                        "sensitive_data_types": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "Types of sensitive data handled",
-                            "default": ["User Data"]
-                        }
-                    },
-                    "required": ["app_description"]
-                }
-            },
-            {
-                "name": "generate_threat_mitigations",
-                "description": "Generate actionable security mitigations for identified threats",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "threats": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Array of threat objects"
-                        },
-                        "priority_filter": {
-                            "type": "string",
-                            "description": "Filter by priority",
-                            "default": "all"
-                        }
-                    },
-                    "required": ["threats"]
-                }
-            },
-            {
-                "name": "create_threat_attack_trees",
-                "description": "Generate application-wide attack tree showing common attack vectors",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "threats": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Array of threat objects (used for context)"
-                        },
-                        "max_depth": {
-                            "type": "integer",
-                            "description": "Maximum tree depth",
-                            "default": 3
-                        },
-                        "output_format": {
-                            "type": "string",
-                            "description": "Output format",
-                            "default": "both"
-                        }
-                    },
-                    "required": ["threats"]
-                }
-            },
-            {
-                "name": "calculate_threat_risk_scores",
-                "description": "Calculate DREAD risk scores to prioritize threats by severity",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "threats": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Array of threat objects"
-                        },
-                        "scoring_guidance": {
-                            "type": "object",
-                            "additionalProperties": True,
-                            "description": "Optional guidance for scoring adjustments"
-                        }
-                    },
-                    "required": ["threats"]
-                }
-            },
-            {
-                "name": "generate_security_tests",
-                "description": "Generate security test cases to validate threat mitigations",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "threats": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Array of threat objects"
-                        },
-                        "test_type": {
-                            "type": "string",
-                            "description": "Type of tests",
-                            "default": "mixed"
-                        },
-                        "format_type": {
-                            "type": "string",
-                            "description": "Output format",
-                            "default": "gherkin"
-                        }
-                    },
-                    "required": ["threats"]
-                }
-            },
-            {
-                "name": "generate_threat_report",
-                "description": "Format complete threat analysis as professional markdown report",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "threat_model": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Array of threat objects"
-                        },
-                        "mitigations": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Optional array of mitigation strategies"
-                        },
-                        "dread_scores": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Optional array of DREAD scores"
-                        },
-                        "attack_trees": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Optional array of attack trees"
-                        },
-                        "include_sections": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "Sections to include in report",
-                            "default": ["executive_summary", "threats", "mitigations", "risk_scores"]
-                        }
-                    },
-                    "required": ["threat_model"]
-                }
-            },
-            {
-                "name": "validate_threat_coverage",
-                "description": "Validate STRIDE coverage completeness and suggest threat model enhancements",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "threat_model": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "additionalProperties": True
-                            },
-                            "description": "Array of threat objects to validate"
-                        },
-                        "app_context": {
-                            "type": "object",
-                            "additionalProperties": True,
-                            "description": "Application context information"
-                        }
-                    },
-                    "required": ["threat_model", "app_context"]
-                }
-            },
-            {
-                "name": "get_repository_analysis_guide",
-                "description": "Get structured framework for extracting threat modeling inputs from repository analysis using GitHub MCP or similar tools",
-                "inputSchema": {
-                    "type": "object",
-                    "properties": {
-                        "analysis_stage": {
-                            "type": "string",
-                            "description": "Analysis stage: 'initial' (quick scan), 'deep_dive' (detailed security analysis), or 'validation' (readiness check)",
-                            "enum": ["initial", "deep_dive", "validation"],
-                            "default": "initial"
-                        },
-                        "repository_context": {
-                            "type": "object",
-                            "description": "Optional context about the repository",
-                            "properties": {
-                                "primary_language": {
-                                    "type": "string",
-                                    "description": "Primary programming language detected"
-                                },
-                                "framework_detected": {
-                                    "type": "string",
-                                    "description": "Primary framework or platform detected"
-                                },
-                                "repository_type": {
-                                    "type": "string",
-                                    "description": "Type of repository",
-                                    "enum": ["application", "library", "infrastructure", "unknown"]
-                                }
-                            }
-                        }
-                    },
-                    "required": []
-                }
-            }
-        ]
-        return {
-            "jsonrpc": "2.0",
-            "result": {"tools": tools},
-            "id": request_id
-        }
-    
-    # Handle tools/call - actual implementation
-    elif method == 'tools/call':
-        tool_name = params.get('name')
-        arguments = params.get('arguments', {})
-        
-        try:
-            if tool_name == 'get_stride_threat_framework':
-                result = get_stride_threat_framework(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-            
-            elif tool_name == 'generate_threat_mitigations':
-                result = generate_threat_mitigations(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-            
-            elif tool_name == 'calculate_threat_risk_scores':
-                result = calculate_threat_risk_scores(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-            
-            elif tool_name == 'create_threat_attack_trees':
-                result = create_threat_attack_trees(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-            
-            elif tool_name == 'generate_security_tests':
-                result = generate_security_tests(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-            
-            elif tool_name == 'generate_threat_report':
-                result = generate_threat_report(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": result  # This returns markdown directly
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-            
-            elif tool_name == 'validate_threat_coverage':
-                result = validate_threat_coverage(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-
-            elif tool_name == 'get_repository_analysis_guide':
-                result = get_repository_analysis_guide(arguments)
-                return {
-                    "jsonrpc": "2.0",
-                    "result": {
-                        "content": [
-                            {
-                                "type": "text",
-                                "text": json.dumps(result, indent=2)
-                            }
-                        ]
-                    },
-                    "id": request_id
-                }
-
-            else:
-                return {
-                    "jsonrpc": "2.0",
-                    "error": {
-                        "code": ERROR_CODES['INVALID_PARAMETER'],
-                        "message": f"Unknown tool: {tool_name}"
-                    },
-                    "id": request_id
-                }
-                
-        except Exception as e:
-            error_id, sanitized_message = sanitize_error(e, f"Tool execution: {tool_name}")
-            return {
-                "jsonrpc": "2.0",
-                "error": {
-                    "code": ERROR_CODES['TOOL_EXECUTION_FAILED'],
-                    "message": sanitized_message
-                },
-                "id": request_id
-            }
-    
-    else:
-        return {
-            "jsonrpc": "2.0",
-            "error": {
-                "code": -32601,
-                "message": f"Method not found: {method}"
-            },
-            "id": request_id
-        }
-
-
-class handler(BaseHTTPRequestHandler):
-    def do_OPTIONS(self):
-        self.send_response(200)
-        self.send_header('Access-Control-Allow-Origin', '*')
-        self.send_header('Access-Control-Allow-Methods', 'GET, POST, OPTIONS')
-        self.send_header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Mcp-Session-Id')
-        self.send_header('X-Content-Type-Options', 'nosniff')
-        self.send_header('X-Frame-Options', 'DENY')
-        self.send_header('X-XSS-Protection', '1; mode=block')
-        self.end_headers()
-        
-    def do_GET(self):
-        self.send_response(200)
-        self.send_header('Content-Type', 'application/json')
-        self.send_header('Access-Control-Allow-Origin', '*')
-        self.send_header('X-Content-Type-Options', 'nosniff')
-        self.send_header('X-Frame-Options', 'DENY')
-        self.send_header('X-XSS-Protection', '1; mode=block')
-        self.end_headers()
-        
-        response_data = {
-            "name": "STRIDE GPT MCP Server",
-            "version": "0.1.0",
-            "description": "Professional threat modeling server using the STRIDE methodology",
-            "tools": [
-                "get_stride_threat_framework",
-                "generate_threat_mitigations",
-                "create_threat_attack_trees",
-                "calculate_threat_risk_scores",
-                "generate_security_tests",
-                "generate_threat_report",
-                "validate_threat_coverage",
-                "get_repository_analysis_guide"
-            ],
-            "endpoints": {
-                "POST /": "MCP JSON-RPC endpoint"
-            }
-        }
-        self.wfile.write(json.dumps(response_data, indent=2).encode())
-        
-    def do_POST(self):
-        try:
-            content_length = int(self.headers.get('Content-Length', 0))
-
-            # Validate payload size before reading
-            if content_length > PAYLOAD_LIMITS['MAX_PAYLOAD_SIZE']:
-                self.send_error_response(413, {
-                    "jsonrpc": "2.0",
-                    "error": {
-                        "code": ERROR_CODES['PAYLOAD_TOO_LARGE'],
-                        "message": f"Payload size {content_length} bytes exceeds maximum of {PAYLOAD_LIMITS['MAX_PAYLOAD_SIZE']} bytes"
-                    },
-                    "id": None
-                })
-                return
-
-            post_data = self.rfile.read(content_length)
-
-            try:
-                body = json.loads(post_data.decode('utf-8'))
-            except json.JSONDecodeError:
-                self.send_error_response(400, {
-                    "jsonrpc": "2.0",
-                    "error": {"code": -32700, "message": "Parse error"},
-                    "id": None
-                })
-                return
-
-            # Validate JSON complexity
-            complexity_result = validate_json_complexity(body)
-            if not complexity_result['valid']:
-                self.send_error_response(400, {
-                    "jsonrpc": "2.0",
-                    "error": {
-                        "code": ERROR_CODES['PAYLOAD_TOO_COMPLEX'],
-                        "message": f"Payload complexity validation failed: {complexity_result['error']}"
-                    },
-                    "id": body.get('id')
-                })
-                return
-
-            # Validate JSON-RPC
-            if not body.get('jsonrpc') == '2.0' or not body.get('method'):
-                self.send_error_response(400, {
-                    "jsonrpc": "2.0",
-                    "error": {"code": -32600, "message": "Invalid Request"},
-                    "id": body.get('id')
-                })
-                return
-            
-            # Handle MCP request using our improved server
-            response = handle_mcp_request(body)
-            
-            self.send_response(200)
-            self.send_header('Content-Type', 'application/json')
-            self.send_header('Access-Control-Allow-Origin', '*')
-            self.send_header('X-Content-Type-Options', 'nosniff')
-            self.send_header('X-Frame-Options', 'DENY')
-            self.send_header('X-XSS-Protection', '1; mode=block')
-            self.end_headers()
-            self.wfile.write(json.dumps(response).encode())
-            
-        except Exception as e:
-            error_id, sanitized_message = sanitize_error(e, "HTTP POST request handling")
-            self.send_error_response(500, {
-                "jsonrpc": "2.0",
-                "error": {"code": -32603, "message": sanitized_message},
-                "id": None
-            })
-    
-    def send_error_response(self, status_code, error_data):
-        self.send_response(status_code)
-        self.send_header('Content-Type', 'application/json')
-        self.send_header('Access-Control-Allow-Origin', '*')
-        self.end_headers()
-        self.wfile.write(json.dumps(error_data).encode())
